Delete role member using delete api.
